requirements_dev.txt
coveralls==1.2.0

selenium==3.9.0
splinter==0.7.7

pytest-django==3.1.2
pytest-splinter==1.8.5
pytest-cov==2.5.1
pytest-pythonpath==0.7.1

mock==2.0.0
model_mommy==1.5.1
psycopg2==2.7.1

bumpversion==0.5.3
tox==2.9.1

